Les identifiants des ressources ne sont plus définies statiquement dans ResourceDescriptor.xml mais autogénérés par une fonction de hashage executée sur des variables membres de la classe Resource

+++ src/fmi/dndtabletop/ihm/CreateBattlefieldDialog.java
@@ -67,7 +67,7 @@ public class CreateBattlefieldDialog extends JDialog {
 		panTile.setLayout(new GridLayout(1, 1));
 		ResourceManager resMan = ResourceManager.getInstance();
 		Vector<Resource> vResources = new Vector<Resource>();
-		for (Map.Entry<Integer, TileResource> entry : resMan.getTextureTileMap().entrySet())
+		for (Map.Entry<Long, TileResource> entry : resMan.getTextureTileMap().entrySet())
 		{
 			vResources.add(entry.getValue());
 		}
