+++ src/fmi/dndtabletop/model/Battlefield.java
@@ -16,7 +16,7 @@ public class Battlefield implements Serializable{
 
 	private static final long serialVersionUID = 1L;
 	
-	public Battlefield(int width, int height, int tileId)
+	public Battlefield(int width, int height, long tileId)
 	{
 		m_width = width;
 		m_height = height;
